pyproject.toml
[project]
name = "mcp-docx"
version = "0.2.0"
description = "MCP server for Word documents: generate, edit, and manipulate .docx from Cursor/IDE"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "mcp[cli]>=1.0.0",
    "python-docx>=1.0.0",
    "docxtpl>=0.16.0",
]

[project.optional-dependencies]
images = ["Pillow>=9.0.0"]
pdf = ["pdf2docx>=0.5.0"]
html = ["mammoth>=1.6.0"]
mailmerge = ["docx-mailmerge>=0.5.0"]
pandoc = ["pypandoc>=1.11"]
all = [
    "Pillow>=9.0.0",
    "pdf2docx>=0.5.0",
    "mammoth>=1.6.0",
    "docx-mailmerge>=0.5.0",
    "pypandoc>=1.11",
]

[project.scripts]
mcp-docx = "mcp_docx.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_docx"]
